requirements.txt
python-telegram-bot==21.*
pandas
rapidfuzz
pydantic
python-dotenv
loguru